% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CausalBootstrap.R
\name{CausalBootstrap}
\alias{CausalBootstrap}
\title{Sharp variance estimator and causal bootstrap in stratified randomized experiments}
\usage{
CausalBootstrap(Y, Z, M, pair = FALSE, NB = 2000, alpha = 0.05)
}
\arguments{
\item{Y}{A numeric vector containing the outcome variable for each observation.}

\item{Z}{A binary vector (0/1) indicating treatment assignment for each observation (0 = control, 1 = treatment).}

\item{M}{A grouping variable, where each unique value corresponds to a different stratum (from 1 to L).}

\item{pair}{A logical value indicating whether the design is paired or not. Default is FALSE.}

\item{NB}{An integer specifying the number of bootstrap samples to generate. Default is 2000.}

\item{alpha}{A numeric value representing the significance level for confidence intervals. Default is 0.05.}
}
\value{
A list containing:
\describe{
\item{tauhat}{The estimated average treatment effect across groups.}
\item{var.neyman}{The Neyman variance estimator for the treatment effect.}
\item{ci.neyman}{The Neyman confidence interval for the treatment effect.}
\item{var.sharp}{The sharp variance estimator for the treatment effect.}
\item{ci.sharp}{The confidence interval for the treatment effect using a sharp variance estimator.}
\item{ci.boot}{The confidence interval for the treatment effect obtained from the bootstrap method.}
\item{var.pair}{The variance estimator for the treatment effect (in paired experiments).}
\item{ci.pair}{The confidence interval for the treatment effect (in paired experiments).}
\item{ci.pair.boot}{The confidence interval for the treatment effect obtained from the bootstrap method (in paired experiments).}
}
}
\description{
This function estimates the average treatment effect using a bootstrap method.
It takes into consideration grouping variables and treatment assignments to provide
a robust measure of treatment effects and their confidence intervals.
}
\examples{
# pair = FALSE
CausalBootstrap(Y = test_stratified$Y, Z = test_stratified$Z, M = test_stratified$M)

# pair = TRUE
CausalBootstrap(Y = test_pair$Y, Z = test_pair$Z, M = test_pair$M, pair = TRUE)


}
